Check if shutil is not None in Requests.py. Fixes #5331

+++ src/python/WMCore/Services/Requests.py
@@ -581,4 +581,5 @@ class TempDirectory():
         self.dir = idir
 
     def __del__(self):
-        shutil.rmtree(self.dir, ignore_errors = True)
+        if shutil:
+            shutil.rmtree(self.dir, ignore_errors = True)
